+++ apps/client/src/hoc/withTopic.js
@@ -24,13 +24,11 @@ const withTopic = (Component) => (props) => {
       slug,
       {
         answers: null,
-        finishedQuestions: false,
-        finishedLocation: false,
         questionIndex: 0,
       },
     ]);
 
-    console.warn("Resseting checker, redirecting to intro page");
+    console.warn("Resetting checker, redirecting to intro page");
     return <Redirect to={geturl(routes.intro, topic)} />;
   }