+++ apps/client/src/pages/IntroPage.js
@@ -1,19 +1,28 @@
-import React, { Suspense } from "react";
+import React, { Suspense, useContext } from "react";
 import { Helmet } from "react-helmet";
 import { useHistory } from "react-router-dom";
 
 import Layout from "../components/Layouts/DefaultLayout";
 import Loading from "../components/Loading";
 import Nav from "../components/Nav";
+import { SessionContext } from "../context";
 import withTopic from "../hoc/withTopic";
 import { geturl, routes } from "../routes";
 
 const IntroPage = ({ topic }) => {
+  const sessionContext = useContext(SessionContext);
   const history = useHistory();
-  const { text, intro } = topic;
+  const { text, intro, slug } = topic;
   const Intro = React.lazy(() => import(`../intros/${intro}`));
 
   const goToNext = () => {
+    sessionContext.setSessionData([
+      slug,
+      {
+        activeComponents: ["locationInput"],
+        finishedComponents: [],
+      },
+    ]);
     history.push(geturl(routes.wrapper, topic));
   };
   return (