+++ apps/client/src/pages/WrapperPage.js
@@ -5,8 +5,8 @@ import { ComponentWrapper } from "../atoms";
 import Conclusion from "../components/Conclusion";
 import DebugDecisionTable from "../components/DebugDecisionTable";
 import Layout from "../components/Layouts/DefaultLayout";
-import Address from "../components/Location/Address";
-import Location from "../components/Location/Location";
+import LocationInput from "../components/Location/LocationInput";
+import LocationResult from "../components/Location/LocationResult";
 import Questions from "../components/Questions";
 import {
   StepByStepItem,
@@ -18,16 +18,23 @@ import withChecker from "../hoc/withChecker";
 const WrapperPage = ({ checker, topic, resetChecker }) => {
   const sessionContext = useContext(SessionContext);
   const { slug, sttrFile } = topic;
-  const { activeComponents, finishedComponents } = sessionContext[slug] || []; // At startup we don't have activeComponent or finishedComponents, so start with empty array.
+
+  // At startup we don't have activeComponent or finishedComponents, so start with empty array.
+  const { activeComponents, finishedComponents } = sessionContext[slug];
 
   // Only one component can be active at the same time.
   const setActiveState = (component) => {
     sessionContext.setSessionData([slug, { activeComponents: [component] }]);
   };
 
-  // Pass a component name, or a array of components, this array or component will all be finished or removed.
+  /**
+   * Set given components to the given finished state
+   *
+   * @param { Object[] | string } component - name of names of the components
+   * @param { boolean } value - Set components to finished state or remove from finished state.
+   */
   const setFinishedState = (component, value) => {
-    let newFinishedComponents = finishedComponents || []; // If there are no finishedComponents start with a empty array.
+    let newFinishedComponents = finishedComponents; // If there are no finishedComponents start with a empty array.
     const allComponents = Array.isArray(component) ? component : [component];
     if (typeof value === "boolean" && value === false) {
       newFinishedComponents =
@@ -35,22 +42,20 @@ const WrapperPage = ({ checker, topic, resetChecker }) => {
     } else {
       newFinishedComponents = [...newFinishedComponents, ...allComponents];
     }
-    // Save the new sessionData to the contex
+    // Save the new sessionData to the context
     sessionContext.setSessionData([
       slug,
       { finishedComponents: newFinishedComponents },
     ]);
   };
 
-  // @TODO: maybe this code can be refactored to one function, but i think this is really declarative. Lets discuss
-  const isActive = (component) => {
+  const isActive = (component, finished = false) => {
+    const components = finished ? finishedComponents : activeComponents;
     const allComponents = Array.isArray(component) ? component : [component];
-    return activeComponents?.includes(...allComponents) || false;
+    return components.includes(...allComponents) || false;
   };
-  const isFinished = (component) => {
-    const allComponents = Array.isArray(component) ? component : [component];
-    return finishedComponents?.includes(...allComponents) || false;
-  };
+
+  const isFinished = (component) => isActive(component, true);
 
   return (
     <Layout>
@@ -67,46 +72,48 @@ const WrapperPage = ({ checker, topic, resetChecker }) => {
             highlightActive
           >
             <StepByStepItem
-              active={
-                (isActive("location") ||
-                  isActive("address") ||
-                  !activeComponents) &&
-                !isFinished("address")
-              }
-              checked={isFinished("address")}
+              active={isActive("locationInput") || isActive("locationResult")}
+              checked={isFinished("locationResult")}
               heading="Adres gegevens"
               largeCircle
             >
-              {(!activeComponents || isActive("location")) && (
-                <Location
-                  topic={topic}
-                  resetChecker={resetChecker}
-                  setActiveState={setActiveState}
+              {(!activeComponents || isActive("locationInput")) && (
+                <LocationInput
+                  {...{
+                    topic,
+                    resetChecker,
+                    setActiveState,
+                  }}
                 />
               )}
-              {(isActive("address") || isFinished("address")) && (
-                <Address
-                  topic={topic}
-                  isFinished={isFinished}
-                  setActiveState={setActiveState}
-                  setFinishedState={setFinishedState}
+              {(isActive("locationResult") || isFinished("locationResult")) && (
+                <LocationResult
+                  {...{
+                    topic,
+                    isFinished,
+                    setActiveState,
+                    setFinishedState,
+                  }}
                 />
               )}
             </StepByStepItem>
             <StepByStepItem
               checked={isFinished("questions")}
               customSize
-              done={isFinished("address")}
+              done={isFinished("locationResult")}
               heading="Vragen"
               largeCircle
             />
-            {isFinished("address") && (
+            {isFinished("locationResult") && (
               <Questions
-                checker={checker}
-                topic={topic}
-                setFinishedState={setFinishedState}
-                setActiveState={setActiveState}
-                isActive={isActive}
+                {...{
+                  checker,
+                  topic,
+                  setFinishedState,
+                  setActiveState,
+                  isActive,
+                  isFinished,
+                }}
               />
             )}
             <StepByStepItem
@@ -117,23 +124,29 @@ const WrapperPage = ({ checker, topic, resetChecker }) => {
               largeCircle
             >
               {isFinished("questions") && (
-                <Conclusion checker={checker} topic={topic} />
+                <Conclusion {...{ topic, checker }} />
               )}
             </StepByStepItem>
           </StepByStepNavigation>
         )}
 
         {/* OLO-flow only needs the Location component */}
-        {!sttrFile && (!activeComponents || isActive("location")) && (
-          <Location topic={topic} setActiveState={setActiveState} />
-        )}
-        {!sttrFile && isActive("address") && (
-          <Address
-            topic={topic}
-            isFinished={isFinished}
-            setActiveState={setActiveState}
-            setFinishedState={setFinishedState}
-          />
+        {!sttrFile && (
+          <>
+            {isActive("locationInput") && (
+              <LocationInput {...{ topic, setActiveState }} />
+            )}
+            {isActive("locationResult") && (
+              <LocationResult
+                {...{
+                  topic,
+                  isFinished,
+                  setActiveState,
+                  setFinishedState,
+                }}
+              />
+            )}
+          </>
         )}
       </ComponentWrapper>