+++ apps/client/src/utils/test-ids.js
@@ -9,10 +9,6 @@ export const FOOTER = "footer";
 export const QUESTION_PAGE = "question-form";
 export const QUESTION_ANSWERS = "question-answers";
 
-export const CONCLUSION_PAGE = "conclusion-title";
-
-export const RESULTS_PAGE = "results-form";
-
 export const NEXT_BUTTON = "next-button";
 export const PREV_BUTTON = "prev-button";